jsdoc for variations-repetitions
